ai_service/requirements.txt
fastapi
uvicorn
google-generativeai
psycopg2-binary
sqlalchemy
python-dotenv
pydantic
